Move helper functions to util file
